+++ indico/modules/events/registration/util.py
@@ -605,8 +605,10 @@ def generate_spreadsheet_from_registrations(registrations, regform_items, static
                     registration_dict[key] = dt
                 else:
                     registration_dict[key] = dt.date()
+            elif item.id in data:
+                registration_dict[key] = item.field_impl.render_spreadsheet_data(data[item.id])
             else:
-                registration_dict[key] = data[item.id].friendly_data if item.id in data else ''
+                registration_dict[key] = ''
         for name, (title, fn) in special_item_mapping.items():
             if name not in static_items:
                 continue
@@ -1010,7 +1012,7 @@ def snapshot_registration_data(registration):
             data[field.html_field_name] = {'price': regdata.price, 'data': regdata.data,
                                            'storage_file_id': regdata.storage_file_id,
                                            'is_file_field': field.field_impl.is_file_field,
-                                           'friendly_data': regdata.friendly_data}
+                                           'friendly_data': field.field_impl.get_snapshot_data(regdata)}
     return data